pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "actual-budget-import"
version = "1.0.0"
description = "Import Actual Budget CSV exports to hledger journal format"
requires-python = ">=3.11"
dependencies = ["pyyaml>=6.0"]

[project.scripts]
actual-budget-import = "actual_budget_import.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts   = "-v --tb=short"

[tool.coverage.run]
source = ["src"]
omit   = ["tests/*"]
